calculator: add power operator

+++ app/Main.hs
@@ -1,7 +1,6 @@
 module Main where
 
 import NaturalCalculator
-import Data.Decimal
 import Console
 
 showResult :: (Show a) => Either ParseError (Either EvaluationError a) -> String
@@ -14,7 +13,7 @@ calculatorMain :: IO ()
 calculatorMain = do
   putStrLn "Welcome to the Calculator!"
   putStrLn "Enter an expression to calculate the answer:"
-  interactLines (showResult . (eval :: String -> Either ParseError (Either EvaluationError Decimal)))
+  interactLines (showResult . (eval :: String -> Either ParseError (Either EvaluationError Double)))
 
 main :: IO ()
 main = calculatorMain
